+++ flask_monitoringdashboard/__init__.py
@@ -48,18 +48,32 @@ def bind(app):
 
     import os
     # Only initialize unit test logging when running on Travis.
-    if '/home/travis/build/' in os.getcwd():
-        print('Detected running on Travis.')
+    if 'TRAVIS' in os.environ:
         import datetime
         from flask import request
 
+        @user_app.before_first_request
+        def log_current_version():
+            """
+            Logs the version of the user app that is currently being tested.
+            :return:
+            """
+            home = os.path.expanduser("~")
+            with open(home + '/app_version.log', 'w') as log:
+                log.write(config.version)
+
         @user_app.after_request
-        def after_request(response):
+        def log_endpoint_hit(response):
+            """
+            Add log_endpoint_hit as after_request function that logs the endpoint hits.
+            :param response: the response object that the actual endpoint returns
+            :return: the unchanged response of the original endpoint
+            """
             hit_time_stamp = str(datetime.datetime.utcnow())
             home = os.path.expanduser("~")
-            log = open(home + '/endpoint_hits.log', 'a')
-            log.write('"{}","{}"\n'.format(hit_time_stamp, request.endpoint))
-            log.close()
+            with open(home + '/endpoint_hits.log', 'a') as log:
+                log.write('"{}","{}"\n'.format(hit_time_stamp, request.endpoint))
+
             return response
 
     # Add all route-functions to the blueprint