+++ flask_monitoringdashboard/core/forms/__init__.py
@@ -1,8 +1,9 @@
 from flask_wtf import FlaskForm
 from wtforms import validators, SubmitField, PasswordField, StringField
+
 from .daterange import get_daterange_form
-from .slider import get_slider_form
 from .double_slider import get_double_slider_form
+from .slider import get_slider_form
 
 
 class Login(FlaskForm):
@@ -10,8 +11,3 @@ class Login(FlaskForm):
     name = StringField('Username', [validators.data_required()])
     password = PasswordField('Password', [validators.data_required()])
     submit = SubmitField('Login')
-
-
-class RunTests(FlaskForm):
-    """ Used for serving a login form on /{{ link }}/testmonitor. """
-    submit = SubmitField('Run selected tests')