+++ flask_monitoringdashboard/core/forms/slider.py
@@ -3,13 +3,15 @@ from flask_wtf import FlaskForm
 from wtforms import SubmitField
 from wtforms.fields.html5 import IntegerRangeField
 
+DEFAULT_SLIDER_VALUE = 10
+
 
 class SliderForm(FlaskForm):
     """
         Class for generating a slider that can be used to reduce the graph.
     """
     slider = IntegerRangeField()
-    submit = SubmitField('Submit')
+    submit = SubmitField('Update')
     title = 'Select a number below for reducing the size of the graph'
 
     def get_slider_value(self):
@@ -36,17 +38,20 @@ class SliderForm(FlaskForm):
                            self.submit(class_="btn btn-primary btn-block"))
 
 
-def get_slider_form(slider_max=100):
+def get_slider_form(slider_max=100, title=None):
     """
     Return a SliderForm with the range from 0 to slider_max
     :param slider_max: maximum value for the slider
+    :param title: override the default title
     :return: a SliderForm with the range (0 ... slider_max)
     """
     form = SliderForm(request.form)
     form.min_value = 1
     form.max_value = slider_max
+    if title:
+        form.title = title
     if 'slider' in request.form:
         form.start_value = request.form['slider']
     else:
-        form.start_value = min(max(form.min_value, form.min_value + (slider_max - form.min_value) // 2), form.max_value)
+        form.start_value = min(DEFAULT_SLIDER_VALUE, slider_max)
     return form