+++ flask_monitoringdashboard/database/__init__.py
@@ -29,36 +29,6 @@ class MonitorRule(Base):
     last_accessed = Column(DateTime)
 
 
-class Tests(Base):
-    """ Table for storing which tests to run. """
-    __tablename__ = 'tests'
-    # name must be unique and acts as a primary key
-    name = Column(String(250), primary_key=True)
-    # boolean to determine whether the test should be run
-    run = Column(Boolean, default=True)
-    # the timestamp of the last time the test was run
-    lastRun = Column(DateTime)
-    # whether the test succeeded
-    succeeded = Column(Boolean)
-
-
-class TestRun(Base):
-    """ Table for storing test results. """
-    __tablename__ = 'testRun'
-    # name of executed test
-    name = Column(String(250), primary_key=True)
-    # execution_time in ms
-    execution_time = Column(Float, primary_key=True)
-    # time of adding the result to the database
-    time = Column(DateTime, primary_key=True)
-    # version of the website at the moment of adding the result to the database
-    version = Column(String(100), nullable=False)
-    # number of the test suite execution
-    suite = Column(Integer)
-    # number describing the i-th run of the test within the suite
-    run = Column(Integer)
-
-
 class FunctionCall(Base):
     """ Table for storing measurements of function calls. """
     __tablename__ = 'functionCalls'
@@ -105,8 +75,25 @@ class Outlier(Base):
     time = Column(DateTime)
 
 
+class TestRun(Base):
+    """ Stores unit test performance results obtained from Travis. """
+    __tablename__ = 'testRun'
+    # name of executed test
+    name = Column(String(250), primary_key=True)
+    # execution_time in ms
+    execution_time = Column(Float, primary_key=True)
+    # time of adding the result to the database
+    time = Column(DateTime, primary_key=True)
+    # version of the user app that was tested
+    version = Column(String(100), nullable=False)
+    # number of the test suite execution
+    suite = Column(Integer)
+    # number describing the i-th run of the test within the suite
+    run = Column(Integer)
+
+
 class TestsGrouped(Base):
-    """ Table for storing grouped tests on endpoints. """
+    """ Stores which endpoints are tested by which unit tests. """
     __tablename__ = 'testsGrouped'
     # Name of the endpoint
     endpoint = Column(String(250), primary_key=True)
@@ -144,4 +131,4 @@ def session_scope():
 
 
 def get_tables():
-    return [MonitorRule, Tests, TestRun, FunctionCall, Outlier, TestsGrouped]
+    return [MonitorRule, FunctionCall, Outlier, TestRun, TestsGrouped]