+++ flask_monitoringdashboard/test/views/testmonitor/test_testmonitor.py
@@ -1,0 +1,6 @@
+import unittest
+
+
+class TestCollectPerformance(unittest.TestCase):
+    def test_collect_perf_dummy(self):
+        self.assertEqual(True, True)