+++ flask_monitoringdashboard/views/details/time_version.py
@@ -3,14 +3,16 @@ from flask import render_template
 from flask_monitoringdashboard import blueprint
 from flask_monitoringdashboard.core.auth import secure
 from flask_monitoringdashboard.core.colors import get_color
+from flask_monitoringdashboard.core.forms import get_slider_form
 from flask_monitoringdashboard.core.plot import boxplot, get_figure, get_layout, get_margin
 from flask_monitoringdashboard.core.plot.util import get_information
 from flask_monitoringdashboard.core.utils import get_endpoint_details, simplify
 from flask_monitoringdashboard.database import FunctionCall, session_scope
+from flask_monitoringdashboard.database.count import count_versions_end
 from flask_monitoringdashboard.database.count_group import get_value
 from flask_monitoringdashboard.database.data_grouped import get_version_data_grouped
 from flask_monitoringdashboard.database.endpoint import to_local_datetime
-from flask_monitoringdashboard.database.versions import get_versions, get_first_requests
+from flask_monitoringdashboard.database.versions import get_first_requests
 
 TITLE = 'Per-Version Performance'
 
@@ -25,10 +27,11 @@ graph you can found out whether the performance changes across different version
 @secure
 def versions(end):
     with session_scope() as db_session:
+        form = get_slider_form(count_versions_end(db_session, end), title='Select the number of versions')
         details = get_endpoint_details(db_session, end)
-        graph = versions_graph(db_session, end)
+        graph = versions_graph(db_session, end, form)
         return render_template('fmd_dashboard/graph-details.html', details=details, graph=graph,
-                               title='{} for {}'.format(TITLE, end),
+                               title='{} for {}'.format(TITLE, end), form=form,
                                information=get_information(AXES_INFO, CONTENT_INFO))
 
 
@@ -43,16 +46,15 @@ def format_version(version, first_used):
     return '{}<br>{}'.format(version, to_local_datetime(first_used).strftime('%Y-%m-%d %H:%M'))
 
 
-def versions_graph(db_session, end):
-    versions = get_versions(db_session, end=end)
+def versions_graph(db_session, end, form):
     times = get_version_data_grouped(db_session, lambda x: simplify(x, 10), FunctionCall.endpoint == end)
-    used = get_first_requests(db_session)
-    data = [boxplot(name=format_version(v, get_value(used, v)), values=get_value(times, v),
-                    marker={'color': get_color(v)})
-            for v in versions]
+    first_requests = get_first_requests(db_session, form.get_slider_value())
+    data = [boxplot(name=format_version(request.version, get_value(first_requests, request.version)),
+                    values=get_value(times, request.version), marker={'color': get_color(request.version)})
+            for request in first_requests]
 
     layout = get_layout(
-        height=350 + 40 * len(versions),
+        height=350 + 40 * len(first_requests),
         xaxis={'title': 'Execution time (ms)'},
         yaxis={'type': 'category', 'title': 'Version', 'autorange': 'reversed'},
         margin=get_margin()