+++ docs/changelog.rst
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 Please note that the changes before version 1.10.0 have not been documented.
 
+Unreleased
+----------
+Changed
+
+- Restructuring of Test-Monitoring page
+
+- Identify testRun by Travis build number
+
 
 v1.13.0
 ----------